Implement TraceContextLayer
A TraceContextLayer is added which adds shared state that can be
accessed by any related spans. This enables tail sampling because now
collection layers like opentelemetry will be able to cache data from
individual spans until the trace is complete, and then decide whether or
not to report it.

To do so, the `TraceContextLayer` should be added to the layer stack
after the Registry but before the OpenTelemetryLayer.