more wip
I figured out a design that would let us generally keep a reporting
layer like otel with minimal changes, but unfortunately there's an
ordering issue with Layer#on_close. Generally, it's good if our trace
recording layer is at the top of the layer stack following the registry.
However, we need all of the other Layer#on_close to run before we could
call like a `#report` method or something.

I'm going to take a new approach which is defining a `TraceData` thing
that lives on the span data in the `Context` struct and optionally
cache there until the final layer on_close. Reporting layers like otel
can check for that tracedata struct and utilize it if there, or else
fall back to the current behavior of reporting immediately.

Next steps are basically to fork the open telemetry reporting layer and
implement that instead. My goal is to provide this crate as a PoC with
that reporting layer. Even with this design, it'll be possible to weave
decision making layers throughout the chain and toggle on/off reporting
per layer.