Add tail sampling support to OpenTelemetryLayer
This leverages the trace context to store related spans in a group and a
basic `SampleDecision` that can optionally be added to prevent reporting
a trace. There's room for more nuance here, like being able to ignore
certain spans or span trees, but this is a nice starting place for a
proof-of-concept.