Stab in the dark at fixing memory bloat
It's been a while since I've looked deeply at this, but IIRC, tracing
keeps a cache of spans around, and thus the extensions may also be kept
around, which means the TraceCache may even be kept around. If that's
the case, the list of spans would grow on all of them until it reaches a
maximum value which could be quite a lot of memory. To avoid that, the
allocation is freed both when a trace is sent and in the reset.

note that if this is being kept around, we might need to reset the
`SampleDecision` extension as well.