record error source chain
BREAKING CHANGE: previously error values were recorded using their Debug
representation. They are now reported with their Display implementation. This is
more in line with current best practices for error handling, but we may want
this to be an opt-in behavior.

This is a change in how error values are recorded in the opentelemetry adapter.
For a given field x that contains an error type, record an additional dynamic
field x.chain that contains an array of all errors in the source chain. This
allows users to determine where a high-level error originated.